split crate into "dlpi" and "libdlpi-sys"
Split out the raw library interface into a libdlpi-sys crate that can be
used by other wrappers, and in the process adjust a few of the function
and constant names to match the C source and manual pages.

Also fix run_tests.sh so that it builds the tests as the current user
before escalating to root to run them.